Improve compatibility in complex cross-targeting scenarios.
This allows for a dll compiled against netstandard2.0 with this as a depedency to be consumed from another dll compiled against net6.0 (or greater) without requiring the netstandard2.0 dll to multitarget netstandard2.0 and net6.0.

This is achieved by the removal of the conditional remove of the output dlls in the package, instead opting for conditional exclusion of most of the codebase except for three [assembly: TypeForwardedTo()] declarations which allow the compiler and runtimes to swap out the type decalrations here for the real ones in net6.0.